requirements.txt
atlassian-python-api
customtkinter
python-dotenv
requests
urllib3
langchain-core
langgraph
pillow
beautifulsoup4
python-pptx
